refactor: consolidate payment processing logic and remove deprecated payment_success_v2.py
